pyproject.toml
[project]
name = "etl-bitcoin"
version = "0.1.0"
description = ""
authors = [
    {name = "Joyce",email = "joycemara.sm@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.6"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pylint (>=4.0.2,<5.0.0)"
]
